railway_requirements.txt
beautifulsoup4>=4.14.2
joblib>=1.5.2
lxml>=5.4.0
numpy>=2.3.2
pandas>=2.3.2
plotly>=6.3.0
psycopg2-binary>=2.9.11
python-telegram-bot[callback-data]>=22.5
pytz>=2025.2
requests>=2.32.5
schedule>=1.2.2
scikit-learn>=1.7.2
scipy>=1.16.2
selenium>=4.35.0
streamlit>=1.49.1
streamlit-autorefresh>=1.0.1
telegram>=0.0.1
tensorflow>=2.20.0
trafilatura>=2.0.0
xgboost>=3.0.5
